move Event* classes out of Particle.{h,cxx} into their own files
